+++ scripts/publish.ts
@@ -1,7 +1,7 @@
 import { execSync } from 'child_process'
-import { join } from 'path'
+import { join, sep } from 'path'
 import consola from 'consola'
-import { readJSONSync, writeJSONSync } from '../build/utils/fs'
+import { readJSONSync, writeJSONSync } from '@node-kit/extra.fs'
 import { version } from '../package.json'
 import { packages } from '../build/packages'
 
@@ -16,7 +16,9 @@ if (version.includes('beta')) command += ' --tag beta'
 if (version.includes('alpha')) command += ' --tag alpha'
 
 for (const { name, pkgName } of packages) {
-	const PKG_FILE = join(PACKAGE, name, 'package.json')
+	const dirName = name.replace(/\./g, sep)
+	const cwd = name === 'monorepo' ? ROOT : join(PACKAGE, dirName)
+	const PKG_FILE = join(cwd, 'package.json')
 	const pkgJson = readJSONSync(PKG_FILE)
 	const newPkgJson = JSON.parse(JSON.stringify(pkgJson))
 	for (const { pkgName: pkg } of packages) {
@@ -35,7 +37,7 @@ for (const { name, pkgName } of packages) {
 	})
 	execSync(command, {
 		stdio: 'inherit',
-		cwd: join(PACKAGE, name)
+		cwd
 	})
 	writeJSONSync(PKG_FILE, pkgJson, {
 		encoding: 'utf8'
@@ -46,8 +48,3 @@ for (const { name, pkgName } of packages) {
 	})
 	consola.success(`Published ${pkgName}`)
 }
-execSync(command, {
-	stdio: 'inherit',
-	cwd: ROOT
-})
-consola.success('Published @node-kit/monorepo')